setting redux for user table info
