+++ package.json
@@ -24,6 +24,8 @@
 		"react-redux": "^7.1.0",
 		"react-router-dom": "^5.0.1",
 		"react-scripts": "3.0.1",
+		"redux": "^4.0.4",
+		"redux-thunk": "^2.3.0",
 		"sessions": "0.0.2-7"
 	},
 	"scripts": {
